Update user permissions when changed
